fixed last cpp 08

+++ cpp08/ex01/Span.cpp
@@ -30,28 +30,29 @@ void Span::addNumber(int n)
 	_numbers.push_back(n);
 }
 
-int Span::shortestSpan() const
+unsigned int Span::shortestSpan() const
 {
 	if (_numbers.size() < 2)
 		throw std::runtime_error("Not enough numbers to find a span");
 	std::vector<int> sorted = _numbers;
 	std::sort(sorted.begin(), sorted.end());
-	int minDiff = std::numeric_limits<int>::max();
+	unsigned int minDiff = std::numeric_limits<unsigned int>::max();
 	for (size_t i = 1; i < sorted.size(); ++i)
 	{
-		int diff = sorted[i] - sorted[i - 1];
+		unsigned int diff = static_cast<unsigned int>(sorted[i])
+			- static_cast<unsigned int>(sorted[i - 1]);
 		if (diff < minDiff)
 			minDiff = diff;
 	}
 	return minDiff;
 }
 
-int Span::longestSpan() const
+unsigned int Span::longestSpan() const
 {
 	if (_numbers.size() < 2)
 		throw std::runtime_error("Not enough numbers to find a span");
 	int minVal = *std::min_element(_numbers.begin(), _numbers.end());
 	int maxVal = *std::max_element(_numbers.begin(), _numbers.end());
-	return maxVal - minVal;
+	return static_cast<unsigned int>(maxVal) - static_cast<unsigned int>(minVal);
 }
